requirements.txt
# Substrate MCP Server Requirements
fastmcp>=0.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
